Remove existing auth logic
The existing logic was based on calls to auth API endpoints and
managing tokens. The redirections on page load was also based on
local storage keys. We are moving away from this structure as we
will use firebase client side authentication methods, at least in
the initial phase. These methods are no longer useful.